requirements.txt
pyupbit==0.2.34
pandas==2.2.3
matplotlib==3.10.1
mplfinance==0.12.10b0
python-dotenv==1.1.0
python-telegram-bot==20.8
anyio==4.9.0
APScheduler==3.6.3
cachetools==4.2.2
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
exceptiongroup==1.2.2
fonttools==4.56.0
h11==0.14.0
httpcore==1.0.7
httpx==0.26.0
idna==3.10
kiwisolver==1.4.8
numpy==1.26.4
packaging==24.2
pillow==9.5.0
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
scipy==1.12.0
six==1.17.0
sniffio==1.3.1
tornado==6.4.2
typing_extensions==4.13.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.3.0
websockets==15.0.1
koreanize-matplotlib==0.1.1